+++ CHANGES.rst
@@ -43,6 +43,7 @@ Changelog
   [aduchene]
 
 2.2.1 (2025-09-10)
+2.2.0 (unreleased)
 ------------------
 
 - Integrate collective.documentgenerator to generate a documents from content types.
@@ -77,6 +78,7 @@ Changelog
   [aduchene]
 
 2.1.11 (2025-06-19)
+2.1.11 (unreleased)
 -------------------
 
 - Fix an issue with `publication.EditForm`.